// api/types/paths/orders/v0/orders/{orderId}.types.ts
// This code was automatically generated from an OpenAPI description.
// Do not edit this file. Edit the OpenAPI file instead.
// For more information, see https://github.com/pmcelhaney/counterfact/blob/main/docs/faq-generated-code.md

import type { WideOperationArgument } from '../../../../../types.ts';
import type { OmitValueWhenNever } from '../../../../../types.ts';
import type { COUNTERFACT_RESPONSE } from '../../../../../types.ts';
import type { Context } from '../../../../../routes/_.context.ts';
import type { ResponseBuilderFactory } from '../../../../../types.ts';
import type { GetOrderResponse } from '../../../../definitions/GetOrderResponse.js';

export type HTTP_GET = (
  $: OmitValueWhenNever<{
    query: never;
    path: { orderId: string };
    header: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
      400: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
      403: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
      404: {
        headers: {
          'x-amzn-RateLimit-Limit': { schema: string };
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
      429: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
      500: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
      503: {
        headers: {
          'x-amzn-RequestId': { schema: string };
        };
        requiredHeaders: never;
        content: {
          'application/json': {
            schema: GetOrderResponse;
          };
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
  }>
) =>
  | {
      status: 200;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | {
      status: 400;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | {
      status: 403;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | {
      status: 404;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | {
      status: 429;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | {
      status: 500;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | {
      status: 503;
      contentType?: 'application/json';
      body?: GetOrderResponse;
    }
  | { status: 415; contentType: 'text/plain'; body: string }
  | COUNTERFACT_RESPONSE
  | { ALL_REMAINING_HEADERS_ARE_OPTIONAL: COUNTERFACT_RESPONSE };